clippy::unnecessary_cast clippy::useless_conversion clippy::let_and_return

+++ src/spa/pod/choice.rs
@@ -170,7 +170,7 @@ where
                 type_: choice_type.raw,
                 flags,
                 child: spa_pod {
-                    size: child_size as u32,
+                    size: child_size,
                     type_: child_type.raw,
                 },
             },
